pyproject.toml
[tool.poetry]
name = "pywolf"
version = "0.3.3"
description = "python utils"
authors = ["winglechen <winglechen@gmail.com>"]
license = "MIT License"
packages = [
    {include = "pywolf"},
    {include = "bin/**/*.py"},
]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.2.1"
boltons = "^23.0.0"
pyyaml = "^6.0"
invoke = "^2.0.0"
aiomysql = "^0.1.1"
databases = "^0.7.0"
httpx = "^0.23.3"
pytest-asyncio = "^0.21.0"
redis = "^4.5.4"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
